* Fix player hype to double and accumulate during gameplay

+++ include/TINS2026/GameplayProgress.hpp
@@ -18,7 +18,7 @@ namespace TINS2026
    public:
       std::set<uint32_t> items_collected_tmj_ids;
       AllegroFlare::Vec2D player_location;
-      int player_excitement;
+      double player_excitement;
       int player_line_capacity;
       int player_lines_filled;
       std::set<uint32_t> progress_and_state_flags;
